workers/3_finalization/pdf_generator/requirements.txt
boto3
fpdf2
Pillow
